+++ CHANGES.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- (BREAKING) Prohibit `AmazingActivist::Base#new` calls. All activities must be
+  called via `.call`, i.e. `A.call(...)` instead of `A.new(...).call`.
+
 
 ## [0.7.0] - 2025-04-14